# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tourist",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "localhost",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Autentificación de usuarios",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@gmail.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/auth/info": {
            "get": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Información del usuario autenticado",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::info",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/recuperar/{email}": {
            "get": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Recuperar la contraseña usuario",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::recuperarContrasena",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Correo electrónico",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/comentarios/list": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Listado de comentarios",
                "operationId": "App\\Http\\Controllers\\API\\ComentariosController::list_comentarios",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/comentarios/add": {
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Adiciona un comentario",
                "operationId": "App\\Http\\Controllers\\API\\ComentariosController::add_comentario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "texto": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "texto": "comentario"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "login",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}